代码提交

+++ base/dao/src/main/java/com/lvzu/dao/FollowerDao.java
@@ -19,7 +19,7 @@ public interface FollowerDao extends BaseDao<User> {
      * @param followerId 被关注者ID
      * @return 影响条数
      */
-    public int insert(int userId, int followerId,int status);
+    public int insert(long userId, long followerId,int status);
 
     /**
      * 删除一个关注
@@ -27,21 +27,21 @@ public interface FollowerDao extends BaseDao<User> {
      * @param followerId 被关注者ID
      * @return 影响条数
      */
-    public int delete(int userId, int followerId);
+    public int delete(long userId, long followerId);
 
     /**
      * 根据用户ID查找所有该ID的关注者
      * @param userId 用户ID
      * @return 关注者列表
      */
-    public Page<User> selectFollowers(Page<User> page, int userId);
+    public Page<User> selectFollowers(Page<User> page, long userId);
 
     /**
      * 根据关注者ID查找其所关注的用户
      * @param followerId 关注者ID
      * @return 关注列表
      */
-    public Page<User> selectUsers(Page<User> page, int followerId);
+    public Page<User> selectUsers(Page<User> page, long followerId);
 
     /**
      * 取消关注
@@ -51,5 +51,5 @@ public interface FollowerDao extends BaseDao<User> {
      * @param cancelTime 取消关注时间
      * @return 影响条数
      */
-    public int cancelFollow(int userId, int followerId,int status,Date cancelTime);
+    public int cancelFollow(long userId, long followerId,int status,Date cancelTime);
 }
